feat(interfaces): add contact page interfaces
